post code review on testing

+++ app.py
@@ -258,6 +258,7 @@ def profile():
 
 @app.route('/users/delete', methods=["POST"])
 def delete_user():
+    # TODO: update docstring 
     """Delete user."""
 
     if not g.user:
@@ -309,19 +310,16 @@ def messages_show(message_id):
 @app.route('/messages/<int:message_id>/delete', methods=["POST"])
 def messages_destroy(message_id):
     """Delete a message."""
-    print('message delete top')
+    msg = Message.query.get_or_404(message_id)
+    # if not g.user or (msg.user_id is not g.user.id)
     if not g.user:
-        print('no g.user')
         flash("Access unauthorized.", "danger")
         return redirect("/")
 
-
-    msg = Message.query.get(message_id)
-    print('messageid',message_id)
-    # if msg.user_id == g.user.id:
-    print('userid', g.user.id)
-    db.session.delete(msg)
-    db.session.commit()
+    # remove if statement here. If statement is something problematic
+    if msg.user_id == g.user.id:
+        db.session.delete(msg)
+        db.session.commit()
 
     return redirect(f"/users/{g.user.id}")
